+++ apps/meteor/app/livechat/server/api/v1/transcript.ts
@@ -1,11 +1,10 @@
 import { Omnichannel } from '@rocket.chat/core-services';
-import type { IOmnichannelRoom } from '@rocket.chat/core-typings';
+import type { AtLeast, IOmnichannelRoom, IUser } from '@rocket.chat/core-typings';
 import { LivechatRooms, Users } from '@rocket.chat/models';
 import { isPOSTLivechatTranscriptParams, isPOSTLivechatTranscriptRequestParams } from '@rocket.chat/rest-typings';
 
 import { i18n } from '../../../../../server/lib/i18n';
 import { API } from '../../../../api/server';
-import { Livechat } from '../../lib/LivechatTyped';
 import { sendTranscript } from '../../lib/sendTranscript';
 
 API.v1.addRoute(
@@ -66,9 +65,51 @@ API.v1.addRoute(
 				throw new Error('error-invalid-user');
 			}
 
-			await Livechat.requestTranscript({ rid, email, subject, user });
+			await requestTranscript({ rid, email, subject, user });
 
 			return API.v1.success();
 		},
 	},
 );
+
+export async function requestTranscript({
+	rid,
+	email,
+	subject,
+	user,
+}: {
+	rid: string;
+	email: string;
+	subject: string;
+	user: AtLeast<IUser, '_id' | 'username' | 'utcOffset' | 'name'>;
+}) {
+	const room = await LivechatRooms.findOneById(rid, { projection: { _id: 1, open: 1, transcriptRequest: 1 } });
+
+	if (!room?.open) {
+		throw new Meteor.Error('error-invalid-room', 'Invalid room');
+	}
+
+	if (room.transcriptRequest) {
+		throw new Meteor.Error('error-transcript-already-requested', 'Transcript already requested');
+	}
+
+	if (!(await Omnichannel.isWithinMACLimit(room))) {
+		throw new Error('error-mac-limit-reached');
+	}
+
+	const { _id, username, name, utcOffset } = user;
+	const transcriptRequest = {
+		requestedAt: new Date(),
+		requestedBy: {
+			_id,
+			username,
+			name,
+			utcOffset,
+		},
+		email,
+		subject,
+	};
+
+	await LivechatRooms.setEmailTranscriptRequestedByRoomId(rid, transcriptRequest);
+	return true;
+}
